requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
websockets

scikit-learn
pandas
